вынес модели вопросов за пределы основного класса ViewController
