+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,12 @@
+
+//  Created by Vitaly Wexler on 16.08.2024.
+//
+
+import Foundation
+
+
+struct QuizQuestion {
+    let image: String
+    let text: String
+    let correctAnswer: Bool
+}
