+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,13 @@
+
+//  Created by Vitaly Wexler on 16.08.2024.
+
+import Foundation
+
+struct QuizResultsViewModel {
+  // строка с заголовком алерта
+  let title: String
+  // строка с текстом о количестве набранных очков
+  let text: String
+  // текст для кнопки алерта
+  let buttonText: String
+}
