+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,10 @@
+
+//  Created by Vitaly Wexler on 16.08.2024.
+import UIKit
+import Foundation
+
+struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
